+++ CHANGELOG.md
@@ -5,6 +5,11 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 
 # Changelog
 
+## [3.9.2]
+### Fixed
+- Automatically create templates for each active language for email translations.
+- Use the customer's order language when generating the payment link.
+
 ## [3.9.1]
 ### Added
 - Add support for generating multiple payment links.